Reduce the number of calls to btr_cur_optimistic_insert()
Summary:
During the record insertion, InnoDB finds out the page the record
should be inserted, acquires the latch for that page and attempts
to insert the record to the page. If there is not enough room in
the page then the b-tree structure must be changed so it releases
the page latch and then gets the b-tree latch. It then re-computes
the page to which the record should be inserted. Most of the time
this second page is the same page as the previously found page and
it still doesn't have room for the record insertion. Rarely,
another thread (eg. purge thread) gets a hold of the previously
found page and deletes some records on that page so that there is
room for insertion in InnoDB's second attempt.

This diff eliminates the unnecessary attempts to insert the record
into the page by checking if the page has been modified since the
last time the insertion had been attempted.

Test Plan:
I added innodb_optimistic_insert_race test which executes two threads
    an inserter and a deleter. The inserter thread inserts records in
    random order while the deleter deletes in random order. There is
    a debug-only status variable called 'innodb_optimistic_insert_calls_in_pessimistic_descent'
    which counts the number of times btr_cur_optimistic_insert() needed to
    be called during the pessimistic descent. This number is around 8
    for the numbers hardcoded in this test which is much less than
    the number of total pessimistic descents (which is in thousands).

Reviewers: jtolmer

Reviewed By: jtolmer