Attempt to fix rare failure in main.connection_recycle
Summary:
This test failed last night because the stat for
connections closed via recycling was too low. My best
educated guess as to the cause was that the low setting
of wait_timeout combined with a slow running test (test
took 14 seconds to finish) meant that a connection was
closed via wait_timeout rather than recycling. Change
the test so it only runs with a low value of wait_timeout
during the portion of the test targetted at that feature.

Test Plan: jenkins

Reviewers: steaphan

Reviewed By: steaphan