Backport fix for race described in http://bugs.mysql.com/bug.php?id=64284
Summary:
This fixes a race in the DROP TABLE fix. It is possible for an assert
to be raised when insert buffer background reads are done on tablespaces
for recently dropped tablespaces

This backports revision 3876 from the MySQL 5.5 branch. I tested it with a custom
test case that is too invasive to be an mtr test. Fortunately, the test
innodb_plugin.innodb_drop_table_ibuf had intermittent failures from this bug.

Test Plan:
mtr, repeated custom test case

Reviewers: steaphan

Reviewed By: steaphan