Avoid deadly embrace between rename table and insert buffer merge reads.
Summary:
This fixes Oracle bug 14556349. Shane provided the testcase and
I provided the code fix. When a thread doing a rename table sets
space->stop_ios, then background threads doing insert buffer merges will
get blocked in fil_mutex_enter_and_prepare_for_io and won't be able to
finish their work to decrement counters on which the rename thread is
blocked (it is in a sleep & retry loop).

The problem was reproduced and the fix confirmed in a custom test case
that isn't feasible to be an mtr test.

Test Plan:
mtr

Reviewers: steaphan

Reviewed By: steaphan