Fix CREATE TABLE ... LIKE to handle KEY_BLOCK_SIZE
Summary:
When doing a CREATE TABLE ... LIKE where the origin table uses both
ROW_FORMAT=COMPRESSED and a KEY_BLOCK_SIZE which is not the default,
then the created table claims that it is using the same key block as
the origin but in fact uses the default.

SHOW CREATE TABLE will report an incorrect key block size but
innochecksum will show that the file on disk is actually using
the default.

I entered this as http://bugs.mysql.com/bug.php?id=66738

Test Plan: Wrote a new test

Reviewers: nizamordulu

Reviewed By: nizamordulu