Preserve the CLIENT_REMEMBER_OPTIONS flag for compressed connections
Summary:
I heard someone recently state that we turn on compression for the
slaves' connections to their masters. This reminded me that there is
a bug in the bit math for turning on this setting which caused problems
in my past job. Fix it here too.

Specifically, keep the CLIENT_REMEMBER_OPTIONS bit set in the flags for
the IO thread's connection so that on failure connecting to the master
mysql_real_connect does not call mysql_close_free_options and stomp the
connection timeout setting.

Test Plan: mtr

Reviewers: steaphan

Reviewed By: steaphan