Make InnoDB call fsync() periodically when creating a new file.
Summary:
When creating a new file (like the log file or the file for a new table),
InnoDB flushes the write buffer to the disk after writing the entire
file. This causes problems on SSD disks as desribed in the related
Facebook task.

This revision adds a new global variable called innodb_fsync_freq which
is the number of bytes that determines the size of the write buffer after
which innodb will flush the contents of the buffer to disk. The default
value for this variable is 128M.

Test Plan:
On a cleanly stopped server, remove the log files and start mysqld. Put
a breakpoint at the newly-added call to os_file_flush() in os_file_set_size()
to make sure that os_file_flush() is called periodically as expected.

Reviewers: jtolmer

Reviewed By: jtolmer