Throttle the fsyncs caused by the slave IO thread.
Summary:
This revision introduces a new sysvar called relay_fsync_freq
which controls the frequency of fsyncs() by the slave IO thread.
The relay logs are flushed/fsync'ed after each time this many bytes are
written. The default value for this variable is 128M.

Test Plan:
make sure the new code gets executed by running it on a
slave and printing messages and adding debug points.

Reviewers: jtolmer

Reviewed By: jtolmer