Revert connection acceleration again.
Summary:
Revert "Adding back connection acceleration changes."

This reverts the re-commit of the previously reverted changes.
Clarity: this REMOVES these optimizations again.

This change set also causes a deadlock.

This whole change set needs to be completely re-written.
It is being reverted again, and should not be re-added.

Or, this can just wait until 5.6, which has much of this.

Test Plan: Clean revert.  OLM stress tests found this.

Reviewers: zshao

Reviewed By: zshao