Transactional replication gets stricter version handling
Summary:
My first attempt at handling upgrade / downgrade
scenarios had an assumption that the bin log names would
be a continuous sequence. Building new slaves from existing
servers means that assumption is not valid. So, instead of
checking that the prepared slot is less than the committed
slot, change the magic number and do not use the prepared
slot if the magic number of the committed slot is old.

Test Plan: mtr

Reviewers: steaphan

Reviewed By: steaphan