If a connection is killed while reading undo records, free resources.
Summary:
An is-connection-killed check was added to the code which builds
the previous version of a row. When that kill check is hit it
failed to free resources, including purge_sys->latch, which then
causes the server to deadlock. Fix the code to release the resources.

Test Plan: Test to be added in subsequent diff

Reviewers: nizamordulu, steaphan

Reviewed By: steaphan