Do not decompress the compressed-only pages of a table being dropped
Summary:
InnoDB goes through the in-memory pages of the to-be-dropped compressed table
in order to remove the table's records from adaptive hash index. While doing so
it decompresses those pages whose uncompressed copy is not in the memory.

This is wasteful and unnecessary because if the uncompressed copy of a page
is not in the memory then its records can not be in AHI. The uncompressed copy
of the page is removed by buf_LRU_free_block() and it already calls
btr_search_drop_page_hash_index() to remove all the records of the page from
AHI.

If a page is compressed-only, then buf_page_get_gen() can not return a buf_block_t*
without doing too much work so I decided to just return NULL. From the perspective
of the caller this should not make any difference because if the page doesn't have
any records in AHI then the caller doesn't do anything with the page.

Test Plan: Create big tables and drop them. Drop big tables on a slave.

Reviewers: mcallaghan, steaphan

Reviewed By: steaphan