Change Default sync_pool_size to 1024
Summary:
Mark found a perf regression in update-heavy workloads with this at 1.

So, this changes the default to 1024.

Test Plan:
Jenkins

This var was originally set to 1024, and was extensively tested at this.
I only changed it to 1 after that testing seemed to show that was ok.
Mark seems to have found a case that shows my original choice was right.

Reviewers: mcallaghan

Reviewed By: mcallaghan