Fix buffer overrun in buddy allocator.
Summary:
My initialization and shutdown of the buddy allocator
had an off-by-one error in the for loops resulting in a buffer
overflow. This doesn't appear to be hurting us in production
because zip_free is the last element of buf_pool_struct and,
when the global buf_pool is allocated, the various aligning
and allocating from cache seem to give enough buffer that the
extra pointer size that we write doesn't cause problems.

Test Plan: Jenkins

Reviewers: nizamordulu

Reviewed By: nizamordulu